make all hashes 64-bit, and refactor VLQs a bit

+++ src/io/lacuna/bifurcan/IEntry.java
@@ -8,11 +8,11 @@ import java.util.function.BiPredicate;
 public interface IEntry<K, V> {
 
   interface WithHash<K, V> extends IEntry<K, V>, Comparable<WithHash<K, V>> {
-    int keyHash();
+    long keyHash();
 
     @Override
     default int compareTo(WithHash<K, V> o) {
-      return Integer.compare(keyHash(), o.keyHash());
+      return Long.compare(keyHash(), o.keyHash());
     }
   }
 
@@ -20,7 +20,7 @@ public interface IEntry<K, V> {
     return new Maps.Entry<>(key, value);
   }
 
-  static <K, V> IEntry.WithHash<K, V> of(int keyHash, K key, V value) {
+  static <K, V> IEntry.WithHash<K, V> of(long keyHash, K key, V value) {
     return new Maps.HashEntry<>(keyHash, key, value);
   }
